Fixed tests. Also removed test cases for uniform samples, as I don't think that belongs in our API atm as is likely a performance-costing optimization due to frequency of appearance (and the better way to handle it is to filter the sampled lists in the first place, if we want to do that.
